requirements.txt
fastapi
uvicorn[standard]
PyPDF2
pdf2docx
python-multipart
jinja2
Pillow
